:truck: api: Move isRelatedToCertification to complementaryCertificationBadgeRepository
