+++ api/src/shared/domain/usecases/index.js
@@ -2,6 +2,7 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
+import * as complementaryCertificationBadgeRepository from '../../../certification/complementary-certification/infrastructure/repositories/complementary-certification-badge-repository.js';
 import * as badgeCriteriaRepository from '../../../evaluation/infrastructure/repositories/badge-criteria-repository.js';
 import * as activityAnswerRepository from '../../../school/infrastructure/repositories/activity-answer-repository.js';
 import * as activityRepository from '../../../school/infrastructure/repositories/activity-repository.js';
@@ -18,6 +19,7 @@ const usecasesWithoutInjectedDependencies = {
 };
 
 const dependencies = {
+  complementaryCertificationBadgeRepository,
   activityAnswerRepository,
   activityRepository,
   assessmentRepository,
