+++ api/tests/certification/complementary-certification/integration/infrastucture/repositories/complementary-certification-badge-repository_test.js
@@ -204,7 +204,9 @@ describe('Integration | Infrastructure | Repository | Certification | Complement
       await databaseBuilder.commit();
 
       // when
-      const results = await complementaryCertificationBadgeRepository.findAttachableBadgesByIds({ ids: [123] });
+      const results = await complementaryCertificationBadgeRepository.findAttachableBadgesByIds({
+        ids: [123],
+      });
 
       // then
       expect(results).to.deep.equal([
@@ -253,7 +255,9 @@ describe('Integration | Infrastructure | Repository | Certification | Complement
       await databaseBuilder.commit();
 
       // when
-      const results = await complementaryCertificationBadgeRepository.findAttachableBadgesByIds({ ids: [123] });
+      const results = await complementaryCertificationBadgeRepository.findAttachableBadgesByIds({
+        ids: [123],
+      });
 
       // then
       expect(results).to.be.empty;
@@ -327,6 +331,59 @@ describe('Integration | Infrastructure | Repository | Certification | Complement
 
         // then
         expect(error).to.be.instanceof(NotFoundError);
+      });
+    });
+  });
+
+  describe('#isRelatedToCertification', function () {
+    describe('when the badge is not acquired', function () {
+      it('should return false', async function () {
+        // given
+        const badgeId = databaseBuilder.factory.buildBadge({ id: 1 }).id;
+        await databaseBuilder.commit();
+
+        // when
+        const isRelatedToCertification =
+          await complementaryCertificationBadgeRepository.isRelatedToCertification(badgeId);
+
+        // then
+        expect(isRelatedToCertification).to.be.false;
+      });
+    });
+
+    describe('when the badge is present in complementary-certification-badges', function () {
+      it('should return true', async function () {
+        // given
+        const badge = databaseBuilder.factory.buildBadge();
+        const complementaryCertificationId = databaseBuilder.factory.buildComplementaryCertification().id;
+        databaseBuilder.factory.buildComplementaryCertificationBadge({
+          badgeId: badge.id,
+          complementaryCertificationId,
+        }).id;
+        await databaseBuilder.commit();
+
+        // when
+        const isRelatedToCertification = await complementaryCertificationBadgeRepository.isRelatedToCertification(
+          badge.id,
+        );
+
+        // then
+        expect(isRelatedToCertification).to.be.true;
+      });
+    });
+
+    describe('when the badge is present in both complementary-certification-badges and complementary-certification-course-results', function () {
+      it('should return true', async function () {
+        // given
+        const badgeId = databaseBuilder.factory.buildBadge().id;
+        databaseBuilder.factory.buildComplementaryCertificationBadge({ complementaryCertificationId: null, badgeId });
+        await databaseBuilder.commit();
+
+        // when
+        const isNotAssociated = await complementaryCertificationBadgeRepository.isRelatedToCertification(badgeId);
+
+        // then
+        expect(isNotAssociated).to.be.true;
       });
     });
   });
