+++ api/tests/shared/integration/infrastructure/repositories/badge-repository_test.js
@@ -279,56 +279,6 @@ describe('Integration | Repository | Badge', function () {
     });
   });
 
-  describe('#isRelatedToCertification', function () {
-    describe('when the badge is not acquired', function () {
-      it('should return false', async function () {
-        // given
-        const badgeId = databaseBuilder.factory.buildBadge({ id: 1 }).id;
-        await databaseBuilder.commit();
-
-        // when
-        const isRelatedToCertification = await badgeRepository.isRelatedToCertification(badgeId);
-
-        // then
-        expect(isRelatedToCertification).to.be.false;
-      });
-    });
-
-    describe('when the badge is present in complementary-certification-badges', function () {
-      it('should return true', async function () {
-        // given
-        const badge = databaseBuilder.factory.buildBadge();
-        const complementaryCertificationId = databaseBuilder.factory.buildComplementaryCertification().id;
-        databaseBuilder.factory.buildComplementaryCertificationBadge({
-          badgeId: badge.id,
-          complementaryCertificationId,
-        }).id;
-        await databaseBuilder.commit();
-
-        // when
-        const isRelatedToCertification = await badgeRepository.isRelatedToCertification(badge.id);
-
-        // then
-        expect(isRelatedToCertification).to.be.true;
-      });
-    });
-
-    describe('when the badge is present in both complementary-certification-badges and complementary-certification-course-results', function () {
-      it('should return true', async function () {
-        // given
-        const badgeId = databaseBuilder.factory.buildBadge().id;
-        databaseBuilder.factory.buildComplementaryCertificationBadge({ complementaryCertificationId: null, badgeId });
-        await databaseBuilder.commit();
-
-        // when
-        const isNotAssociated = await badgeRepository.isRelatedToCertification(badgeId);
-
-        // then
-        expect(isNotAssociated).to.be.true;
-      });
-    });
-  });
-
   describe('#remove', function () {
     describe('when the record to delete is in the table', function () {
       it('should return true when deletion goes well', async function () {
